functions/myspa-etl/requirements.txt
pandas
openpyxl
google-cloud-storage
google-cloud-bigquery
flask
pandas-gbq